refactor(ui): split instances.$id.tsx god object into modules
